backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# PDF parsing
pdfplumber==0.11.0

# LLM - Groq
groq==0.9.0

# NLP - semantic similarity
sentence-transformers==3.0.1

# NLP - keyword extraction
spacy==3.7.5

# Database
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Email
resend==2.2.0

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.8.0
pydantic-settings==2.3.4

# CSV
pandas==2.2.2